% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caching.R
\name{checkExtension}
\alias{checkExtension}
\title{Perform and cache model extension or load it from cache}
\usage{
checkExtension(
  fname,
  original_model,
  mc = 0,
  sample_range = c(10^(2:-1), 0, -10^(-1:2)),
  padjust_method = "bonferroni",
  parallel = TRUE,
  print = TRUE
)
}
\arguments{
\item{fname}{Name for the saved file (without extension)}

\item{parallel}{Kept for backward compatibility, won't be used}

\item{print}{Provided for convenient copy-paste from suggestExtension lines, has no control here as the results will always be cached in a file.}
}
\description{
All parameters except 'fname' correspond to the ones in suggestExtension
}
