% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caching.R
\name{checkRefitModel}
\alias{checkRefitModel}
\title{Load refited model from cache}
\usage{
checkRefitModel(
  fname,
  initial_model,
  fixed_parameters,
  nb_cores = 1,
  inits = 1000
)
}
\arguments{
\item{fixed_parameters}{A list of parameters to fix in the form 'list("param_name"=param_value)'}
}
\description{
Check if a model exists or use the refitModel function to create it. Provides some helpful functions to change just a few parameters
}
