% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caching.R
\name{checkVariableParameters}
\alias{checkVariableParameters}
\title{Perform variable parameters analysis or load a variable model if it exists}
\usage{
checkVariableParameters(vmodel_name, original_modelset, data.stimulation,
  data.variation = c(), nb_cores = 0, max_iterations = 0,
  nb_samples = 100, accuracy = 0.95, method = "geneticlhs",
  notVariable = c())
}
\description{
Perform variable parameters analysis or load a variable model if it exists
}
