% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeLFC.R
\name{compare_replicates}
\alias{compare_replicates}
\title{Plots to compare replicates}
\usage{
compare_replicates(rep1, rep2, rep1_name = "", rep2_name = "")
}
\arguments{
\item{rep1}{Data for replicate 1}

\item{rep2}{Data for replicate 2}

\item{rep1_name}{Name of replicate 1}

\item{rep2_name}{Name of replicate 2}
}
\description{
Compare replicates with readouts as columns and treatments as rownames
}
