% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeLFC.R, R/processing.R
\name{computeLFC}
\alias{computeLFC}
\title{Compute log-fold changes from a MIDAS file}
\usage{
computeLFC(full_datas)

computeLFC(full_datas)
}
\description{
Compute log-fold changes from a MIDAS file

Compute log-fold change
}
