% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDataSet.R
\name{makeDataSet}
\alias{makeDataSet}
\title{Make a dataset to generate a ModelSet with the same perturbations}
\usage{
makeDataSet(...)
}
\description{
TODO fix to actually only have common ones
TODO add the possibility to add the non common as NA
}
