% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPlateMIDAS.R, R/processing.R
\name{plotPlateMIDAS}
\alias{plotPlateMIDAS}
\title{Plot the values in a plate from a MIDAS file with ID:well column}
\usage{
plotPlateMIDAS(midas_file)

plotPlateMIDAS(midas_file)
}
\description{
Plot the values in a plate from a MIDAS file with ID:well column

Plot values in plate layout
}
